requirements.txt
anthropic
mcp[cli]
sqlparse
aiomysql
pyyaml
hypothesis
pytest
pytest-asyncio
